+++ src/CMakeLists.txt
@@ -64,7 +64,7 @@ add_ippl_sources (Ippl.cpp)
 
 add_subdirectory (Communicate)
 if (ENABLE_FFT)
-    set (HEFFTE_LIBRARY Heffte::Heffte)
+    set (HEFFTE_LIBRARY Heffte)
     add_subdirectory (FFT)
 else()
     set (HEFFTE_LIBRARY "")
