+++ src/MaxwellSolvers/Maxwell.h
@@ -15,32 +15,31 @@
 
 namespace ippl {
 
-    template <typename EMField, typename SourceField>
+    template <typename EMField, typename FourField>
     class Maxwell {
     public:
         // typedefs for the different fields and vector fields
-        using typeR = typename SourceField::value_type;
+        using typeR = typename FourField::value_type;
 
         constexpr static unsigned Dim = EMField::dim;
         typedef typename EMField::Mesh_t Mesh_t;
         typedef typename EMField::Centering_t Centering_t;
         typedef Vector<typeR, Dim> Vector_t;
-        typedef Field<Vector_t, Dim, Mesh_t, Centering_t> VectorSourceField_t;
 
         // define type for field layout
         typedef FieldLayout<Dim> FieldLayout_t;
 
         // type for communication buffers
-        using memory_space = typename SourceField::memory_space;
-        using buffer_type  = Communicate::buffer_type<memory_space>;
+        using memory_space = typename FourField::memory_space;
+        //using buffer_type  = Communicate::buffer_type<memory_space>;
 
         /*!
          * Default constructor for Maxwell solvers;
          */
         Maxwell() {}
 
-        Maxwell(SourceField& charge, VectorSourceField_t& current, EMField& E, EMField& B) {
-            setSources(charge, current);
+        Maxwell(FourField& source, EMField& E, EMField& B) {
+            setSource(source);
             setEMFields(E, B);
         }
 
@@ -49,9 +48,8 @@ namespace ippl {
          * @param charge Reference to rho field
          * @param current Reference to J field
          */
-        virtual void setSources(SourceField& charge, VectorSourceField_t& current) {
-            rhoN_mp = &charge;
-            JN_mp   = &current;
+        virtual void setSource(FourField& charge) {
+            source_mp = &charge;
         }
 
         /*!
@@ -73,8 +71,7 @@ namespace ippl {
 
     protected:
         // fields containing reference to charge and current
-        SourceField* rhoN_mp       = nullptr;
-        VectorSourceField_t* JN_mp = nullptr;
+        FourField* source_mp = nullptr;
 
         // E and B fields
         EMField* En_mp = nullptr;
